:art: Changing packages contents and changing Integer support to Long

+++ main/src/tree/NodeTypes.java
@@ -2,6 +2,6 @@ package tree;
 
 public enum NodeTypes {
     STRING,
-    INTEGER,
+    LONG,
     DATE
 }
